Add buttons
This commit adds support for buttons which can run a function when
clicked on. A button can contain either text or an image; image
takes precedence over text if both are defined.

If you define a callback that doesn't exist, then the button does
nothing. See the new test/example, `tests/township.py` to see how
to make callbacks exist. The callback name given in the `on-click`
field of the definition is looked up in the `yamlui.callbacks`
dictionary.

Callbacks must take a single positional argument (the event that
triggered it) and an arbitrary set of keyword arguments.

You can define keyword arguments that are passed to the callback
by setting the `callback-args` dictionary in the yaml definition
of the widget. This should be a dictionary of the form
`arg_name: value`.

Example:

    - object: button
      properties:
        text: New Game
        position: [10, 10]
        image: images/button.png
        on-click: set_state
        hover-effects:
          image: images/button_hover.png
      callback-args:
        new_state: 'new-game-menu'
        old_state: 'main-menu'