Add a `name` property for widgets
This commit adds a `name` property, which can be used in any widget.
This name should be unique within a given yaml file, and can be used
to access the widget quickly using:

    yamlui.trees.get(filepath, {}).get(name)

This allows callbacks to easily modify properties of widgets, such as
changing the text content of a label.