Allow `content-bind` on labels
This is a one-way data binding, to make it easy to show data from
the internal model on the UI.

The value of the `content-bind` key in the yaml definition should
be a key for `yamlui.callbacks`. This function must take no required
arguments, and must return the value to be displayed on the label.

If `content-bind` is set, then `text` is ignored even if it is
present.